chore: put Draft<T> tests in their own file
